// src/app/interfaces/forecast.interface.ts
// export interface Forecast {
//   coord?: Coord;
//   weather?: Weather[];
//   base?: string;
//   main?: Main;
//   visibility?: number;
//   wind?: Wind;
//   clouds?: Clouds;
//   dt?: number;
//   sys?: Sys;
//   timezone?: number;
//   id?: number;
//   name?: string;
//   cod?: number;
// }

//   export interface Coord {
//       lon: number;
//       lat: number;
//   }

//   export interface Weather {
//       id: number;
//       main: string;
//       description: string;
//       icon: string;
//   }

//   export interface Main {
//       temp: number;
//       feelslike: number;
//       tempMin: number;
//       tempMax: number;
//       pressure: number;
//       humidity: number;
//   }

//   export interface Wind {
//       speed: number;
//       deg: number;
//       gust: number;
//   }

//   export interface Clouds {
//       all: number;
//   }

//   export interface Sys {
//       type: number;
//       id: number;
//       country: string;
//       sunrise: number;
//       sunset: number;
//   }


export interface Root {
  coord: Coord;
  weather: Weather[];
  base: string;
  main: Main;
  visibility: number;
  wind: Wind;
  clouds: Clouds;
  dt: number;
  sys: Sys;
  timezone: number;
  id: number;
  name: string;
  cod: number;
}

export interface Coord {
  lon: number;
  lat: number;
}

export interface Weather {
  id: number;
  main: string;
  description: string;
  icon: string;
}

export interface Main {
  temp: number;
  feelsLike: number;
  tempMin: number;
  tempMax: number;
  pressure: number;
  humidity: number;
}

export interface Wind {
  speed: number;
  deg: number;
}

export interface Clouds {
  all: number;
}

export interface Sys {
  type: number;
  id: number;
  country: string;
  sunrise: number;
  sunset: number;
}